requirements.txt
colorama
feedparser
python-dateutil
pytz
python-dotenv